contributed by Tony Kelman: improvements to Matlab recognition in configure and build of Matlab interface
Following Tony's description of this patch:

Overall, I see that configure has a section regarding the Matlab interface,
but much of the time (especially on Windows or Mac) it does not do the job
entirely and users have to make manual modifications to the Matlab interface
Makefile along with mexopts.sh (.bat on Windows). My goal with this patch is
to at least eliminate the need to modify the Makefile, for as many cases as
possible.

The existing configure section for the Matlab interface tries to
automatically look for the Matlab installation location with a "which"
call, but many recent Matlab versions (for Linux) have moved away from adding
the $MATLAB_HOME/bin directory to the system path, and towards adding just a
few symbolic links in /usr/local/bin. So mexext is rarely on the path
anymore, but matlab frequently is - often as a symbolic link. I've changed the
"which" commands to look for matlab accordingly, and in the case of
symbolic links, read where they point to ($MATLAB_HOME/bin/matlab). I'm not sure
whether my `ls -l "$mexpath" | awk '{print $NF}'` is more or less
preferable/portable than using readlink. I also added a few quotes
anywhere $MATLAB_HOME gets used, in case the Matlab path has spaces in it (common
in Windows - could get or require the user to provide the short PROGRA~1
style name, but I don't think it's necessary).

On Windows, Matlab is more frequently on the path. The behavior of
"which" is inconsistent between Cygwin and MSYS for executable names: Cygwin
returns the location without the .exe extension if the input does not end with
.exe, whereas in MSYS "which matlab" returns "/c/path/to/matlab.exe" or
similar.
If you ask for "which matlab.exe" on the other hand, both Cygwin and
MSYS return the location including the .exe at the end. So I've added $EXEEXT
in the "which" command and the sed call that removes /bin/matlab.exe from
the end of the "which" result. This will make no difference for Linux or Mac
users, when $EXEEXT is empty.

On Windows, the mex and mexext utilities are batch files, so we need to
add .bat in order to find them. MSYS can't execute batch files natively, so
I use cmd /c to call mexext.bat for MinGW.

The mex compiler doesn't understand Apple-style flags like "-framework
vecLib." Ray Zimmerman's "Ipopt on Mac OS X" wiki page has a step where
he says to remove "-framework vecLib" from the LIBS line of the Matlab
interface Makefile, and to add it to LDFLAGS in mexopts.sh. I've added a
sed call in contrib/MatlabInterface/src/Makefile.in to remove it from LIBS
automatically. Rather than adding it to LDFLAGS in mexopts.sh, it can be
added to LDFLAGS in Makefile.in. I've added an AC_SUBST variable to do
that, defined in configure by grepping $BLAS_LIBS and $LAPACK_LIBS for
"-framework vecLib". This way the user doesn't have to mess with mexopts.sh as
often.

There can still be problems on Mac since Mathworks hardcodes an expected
Xcode version in $SDKROOT, which may not be installed. That's not
necessarily configure's job to fix, but it could be. I have a
post-configure script that copies the default mexopts.sh from its location in
$MATLAB_HOME/bin, tests whether or not the expected $SDKROOT directory
exists, and makes the necessary modifications if it doesn't.

On Windows, mex doesn't recognize the -cxx flag. I get rid of it
selectively using a new AM_CONDITIONAL, MEX_WINDOWS. I also use the same conditional
to modify the setting of MEX to call Matlab's included Perl on
$(MATLAB_HOME)/bin/mex.pl. There are a few more quotes and a CYGPATH_W
call in there to handle Matlab paths with spaces, and paths returned from
"which" that are in /cygdrive/c/path format. I was a little worried about the
win32 in the path to Perl, but turns out even 64-bit Matlab includes a 32-bit
version of Perl at that path.

I've tested this set of changes pretty extensively on Linux, Mac, Win32,
Win64, Cygwin, MinGW32, MinGW64, with and without Matlab, various
combinations of specifying --with-matlab-home as a configure flag with
different directory formats, or letting configure find Matlab via
"which."
It works as expected, or at least better than the existing version, in
every test I could come up with.

On Windows this patch by itself is not enough to get a functioning mex
file, but the only missing piece should be mexopts.bat generated by Gnumex
(http://sourceforge.net/projects/gnumex/) then modified appropriately.
The post-configure script I have does all this, but I imagine automatically
downloading Gnumex could be problematic due to the differing licenses.
It might be too OS-specific to add as a ThirdParty/Gnumex/get.Gnumex, but I
could contribute the remaining pieces in that form if there's interest. 



git-svn-id: https://projects.coin-or.org/svn/Ipopt/trunk@2111 ef2a96d0-92f8-0310-a3c1-cdaa70078a94